work in progress: refactoring login functionality from frontend to backend
